# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Perinatal Contacts API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /health:
    get:
      summary: Health
      responses:
        "200":
          description: OK
  /meta/types:
    get:
      summary: Institution types
      responses:
        "200":
          description: OK
  /meta/regions:
    get:
      summary: Regions present in DB
      responses:
        "200":
          description: OK
  /meta/stats:
    get:
      summary: Counts
      responses:
        "200":
          description: OK
  /institutions:
    get:
      summary: List institutions
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: region
          schema: { type: string }
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: has_email
          schema: { type: boolean }
        - in: query
          name: has_phone
          schema: { type: boolean }
        - in: query
          name: has_chief
          description: true — только с известным главврачом, false — только без него
          schema: { type: boolean }
        - in: query
          name: nmic_ref
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - in: query
          name: sort
          schema: { type: string, enum: [name, updated_at, region], default: name }
      responses:
        "200":
          description: Paginated list
  /institutions/{id}:
    get:
      summary: Get by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /institutions/{id}/persons:
    get:
      summary: Персоны учреждения (руководство, зав. отделениями)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
        - in: query
          name: min_confidence
          description: Порог достоверности; по умолчанию low скрыт
          schema: { type: string, enum: [high, medium, low], default: medium }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Person" }
                  total: { type: integer }
        "404":
          description: Not found
  /admin/institutions:
    post:
      summary: Create
      security: [{ ApiKeyAuth: [] }]
      responses:
        "201":
          description: Created
  /admin/institutions/{id}:
    patch:
      summary: Update
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
    delete:
      summary: Soft-reject or delete
      security: [{ ApiKeyAuth: [] }]
      responses:
        "204":
          description: Deleted
  /admin/jobs/crawl:
    post:
      summary: Enqueue crawl
      security: [{ ApiKeyAuth: [] }]
      responses:
        "202":
          description: Accepted
  /admin/jobs/enrich:
    post:
      summary: Обогащение ФИО руководства и email с сайтов учреждений
      security: [{ ApiKeyAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                limit: { type: integer, default: 25, minimum: 1, maximum: 500 }
                only_missing_chief: { type: boolean, default: true }
                region: { type: string, nullable: true }
                type: { type: string, nullable: true }
                with_site_only:
                  type: boolean
                  default: false
                  description: только учреждения с уже известным сайтом
                force:
                  type: boolean
                  default: false
                  description: игнорировать недавние неудачные попытки
      responses:
        "202":
          description: Job создан и выполнен синхронно
  /admin/jobs/{job_id}:
    get:
      summary: Статус job любого вида
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
        "404":
          description: Not found
  /admin/persons:
    get:
      summary: Очередь персон для ручной проверки
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - in: query
          name: institution_id
          schema: { type: string, format: uuid }
        - in: query
          name: role
          schema: { type: string, enum: [chief, deputy, pathology_head, head, other] }
        - in: query
          name: confidence
          schema: { type: string, enum: [high, medium, low] }
        - in: query
          name: unverified_only
          schema: { type: boolean, default: false }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
      responses:
        "200":
          description: OK
  /admin/persons/{person_id}:
    patch:
      summary: Правка персоны; verified_manually защищает от перезаписи парсером
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
        "404":
          description: Not found
    delete:
      summary: Удалить персону и пересчитать поля учреждения
      security: [{ ApiKeyAuth: [] }]
      responses:
        "204":
          description: Deleted
  /admin/metrics/completeness:
    get:
      summary: Метрики полноты данных
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
  /admin/mailings/preview:
    post:
      summary: Предпросмотр рассылки с подстановками
      description: |
        Подстановки в subject и body_html: {{name}}, {{city}}, {{region}}, {{chief}}, {{address}}.
        Если ФИО главврача неизвестно, {{chief}} превращается в «коллеги».
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
  /admin/export:
    post:
      summary: Excel export by filter
      security: [{ ApiKeyAuth: [] }]
      responses:
        "202":
          description: Job created
  /admin/export/{job_id}:
    get:
      summary: Export job status / download link
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
  /admin/mailings:
    post:
      summary: Create mailing (dry_run default true)
      security: [{ ApiKeyAuth: [] }]
      responses:
        "202":
          description: Accepted
  /admin/mailings/{id}:
    get:
      summary: Mailing status
      security: [{ ApiKeyAuth: [] }]
      responses:
        "200":
          description: OK
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Institution:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type: { type: string }
        region: { type: string }
        city: { type: string }
        address: { type: string }
        phones:
          type: array
          items: { type: string }
        emails:
          type: array
          items: { type: string }
        website: { type: string, nullable: true }
        chief_physician: { type: string, nullable: true }
        pathology_head: { type: string, nullable: true }
        nmic_ref: { type: string, nullable: true }
        source_url: { type: string }
        verification_status: { type: string }
        updated_at: { type: string, format: date-time }
    Person:
      type: object
      properties:
        id: { type: string, format: uuid }
        institution_id: { type: string, format: uuid }
        full_name: { type: string }
        role: { type: string, enum: [chief, deputy, pathology_head, head, other] }
        position_raw: { type: string, nullable: true }
        department: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        confidence:
          type: string
          enum: [high, medium, low]
          description: high — ФИО и должность в одном блоке на странице руководства
        source_url: { type: string }
        verified_manually: { type: boolean }